% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gocrMain.R
\name{gocrOptionString}
\alias{gocrOptionString}
\title{Option String for gocr Call}
\usage{
gocrOptionString(config)
}
\arguments{
\item{config}{gocr configuration as returned by \code{\link{gocrConfig}}}
}
\description{
Option String for gocr Call
}
