% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gocrMain.R
\name{gocrRun}
\alias{gocrRun}
\title{Run gocr on an Image File}
\usage{
gocrRun(
  config,
  useBatch = TRUE,
  waitForBatch = TRUE,
  opendir = TRUE,
  dbg = TRUE
)
}
\arguments{
\item{config}{gocr configuration as returned by \code{\link{gocrConfig}}}

\item{useBatch}{if TRUE (default), a batch file is written so that the user
can reproduce the call by double-clicking the batch file in the file
explorer (opens when \emph{opendir} is TRUE)}

\item{waitForBatch}{passed to \code{kwb.gocr:::writeAndRunBatchFile}}

\item{opendir}{if TRUE (default), and if \emph{useBatch} is TRUE, the
directory in which the batch file is written, is opened in the Windows
Explorer}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\value{
(only if \emph{waitForBatch} = TRUE): result of OCR as a vector of 
  character representing the recognised lines. The result vector has the 
  attribute \emph{config} containing the configuration used (original config,
  with default values set where needed)
}
\description{
Run gocr on an Image File
}
