% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gocrMain.R
\name{optionValueMode}
\alias{optionValueMode}
\title{Value for Mode Option}
\usage{
optionValueMode(useDatabase = FALSE, layoutAnalysis = FALSE,
  doNotCompare = FALSE, doNotDivide = FALSE, doNotCorrect = FALSE,
  characterPacking = FALSE, extendDatabase = FALSE,
  switchOffEngine = FALSE)
}
\arguments{
\item{useDatabase}{(2) use database to recognize characters which are not
recognized by other algorithms, (early development)}

\item{layoutAnalysis}{(4) switching on layout analysis or zoning
(development)}

\item{doNotCompare}{(8) don't compare unrecognized characters to recognized
one}

\item{doNotDivide}{(16) don't try to divide overlapping characters to two or
three single characters}

\item{doNotCorrect}{(32) don't do context correction}

\item{characterPacking}{(64) character packing, before recognition starts,
similar characters are searched and only one of this characters will be
send to the recognition engine (development)}

\item{extendDatabase}{(128) extend database, prompts user for unidentified
characters and extends the database with users answer (128+2, early
development)}

\item{switchOffEngine}{(256) switch off the recognition engine (makes sense
together with -m 2)}
}
\description{
Value for Mode Option
}
\references{
\url{http://manpages.ubuntu.com/manpages/gutsy/man1/gocr.1.html}
}
