% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gocrMain.R
\name{optionValueVerbosity}
\alias{optionValueVerbosity}
\title{Value for Option Verbosity}
\usage{
optionValueVerbosity(printMore = 1, listShapes = 1, listPattern = 1,
  printPattern = 1, printDebug = 1, createOutPng = 0)
}
\arguments{
\item{printMore}{(1) print more info}

\item{listShapes}{(2) list shapes of boxes (see -c) to stderr}

\item{listPattern}{(4) list pattern of boxes (see -c) to stderr}

\item{printPattern}{(8) print pattern after recognition for debugging}

\item{printDebug}{(16) print debug information about recognition of lines to
stderr}

\item{createOutPng}{(32) create outXX.png with boxes and lines marked on each
general OCR-step}
}
\description{
Value for Option Verbosity
}
